Extract Tab from TabbedResource
