+++ front/app/typings.ts
@@ -2,6 +2,17 @@ import {
   appLocalePairs,
   appGraphqlLocalePairs,
 } from 'containers/App/constants';
+import { Messages } from 'react-intl';
+import { IProjectAction } from 'services/projects';
+import { IIdeaAction } from 'services/ideas';
+import { FormikActions } from 'formik';
+import { FC } from 'react';
+import { TableCellProps } from 'semantic-ui-react';
+import {
+  TAppConfigurationSetting,
+  TAppConfigurationSettingWithEnabled,
+} from 'services/appConfiguration';
+import { TFieldName } from 'components/UI/Error';
 
 declare global {
   interface Function {
@@ -54,7 +65,7 @@ export interface ITab {
   label: string;
   url: string;
   active?: boolean | ((pathname: string) => boolean);
-  feature?: TAppConfigurationSetting;
+  feature?: TAppConfigurationSettingWithEnabled;
   statusLabel?: string;
 }
 
@@ -111,14 +122,6 @@ export interface Message {
   defaultMessage: string;
 }
 
-import { Messages } from 'react-intl';
-import { IProjectAction } from 'services/projects';
-import { IIdeaAction } from 'services/ideas';
-import { FormikActions } from 'formik';
-import { FC } from 'react';
-import { TableCellProps } from 'semantic-ui-react';
-import { TAppConfigurationSetting } from 'services/appConfiguration';
-import { TFieldName } from 'components/UI/Error';
 export type MessageDescriptor = Messages['key'];
 
 export type Locale = keyof typeof appLocalePairs;
